test(gateway): split session files transcript-fold coverage
